Move conversation mapping code into payload class
